Splitting Utils/SourceAssetsStorage into smaller, thematically coherent files.
