+++ Gems/ROS2RobotImporter/Code/Source/RobotImporter/URDF/SdfPrefabMaker.h
@@ -24,7 +24,7 @@
 #include <AzCore/std/string/string.h>
 #include <AzToolsFramework/Prefab/PrefabIdTypes.h>
 #include <AzToolsFramework/Prefab/PrefabPublicInterface.h>
-#include <RobotImporter/Utils/SourceAssetsStorage.h>
+#include <RobotImporter/Assets/AssetTypes.h>
 #include <optional>
 
 namespace ROS2RobotImporter
